Renameing to Home-Assistant LiteDB
